requirements/base.txt
Django==5.1
python-environ==0.4.54
psycopg2-binary==2.9.9
django-phonenumber-field==8.0.0
djangorestframework==3.15.2